+++ src/certificate/index.ts
@@ -2,6 +2,8 @@ export {
   generateCSR,
   generateECDSAKeyPair,
   extractPublicKey,
+  extractRawPublicKey,
+  extractCertificateSignature,
   encryptPrivateKey,
   decryptPrivateKey,
 } from './generate.js';